refactor: Removed dayjs library and replaced date manipulation with native JavaScript Date objects
