+++ package.json
@@ -24,7 +24,6 @@
     "@mui/material": "^6.4.0",
     "@reduxjs/toolkit": "^2.5.0",
     "axios": "^1.7.9",
-    "dayjs": "^1.11.13",
     "formik": "^2.4.6",
     "i18next": "^24.2.2",
     "i18next-browser-languagedetector": "^8.0.2",
