+++ src/components/calendar/Calendar.jsx
@@ -1,16 +1,18 @@
 import FullCalendar from "@fullcalendar/react";
 import dayGridPlugin from "@fullcalendar/daygrid";
 import interactionPlugin from "@fullcalendar/interaction";
-import dayjs from "dayjs";
 
 const Calendar = ({ handleDateSelect }) => {
+  const today = new Date();
+  const todayFormatted = today.toISOString().split("T")[0];
+
   return (
     <FullCalendar
       plugins={[dayGridPlugin, interactionPlugin]}
       initialView="dayGridWeek"
       selectable={true}
       select={(info) => handleDateSelect(info.startStr)}
-      validRange={{ start: dayjs().format("YYYY-MM-DD") }}
+      validRange={{ start: todayFormatted }}
       headerToolbar={{
         left: "prev,next",
         center: "title",
